feat: 메인페이지 sticky 애니메이션
