+++ src/components/HighlightenTitle.tsx
@@ -2,13 +2,16 @@ import { cn } from '@/lib/utils'
 
 export default function HighlightenTitle({
     text,
-    className
+    className,
+    ref
 }: {
     text: string
     className?: string
+    ref?: React.RefObject<HTMLDivElement>
 }) {
     return (
         <h1
+            ref={ref}
             className={cn(
                 'font-pretendard text-pri-white bg-sub-seoultech-red text-64 w-fit leading-none font-bold tracking-[-1.92px]',
                 className
